+++ packages/edition-slices/__tests__/shared-slices.base.js
@@ -5,7 +5,8 @@ import {
   mockLeadTwoNoPicAndTwoSlice,
   mockSecondaryFourSlice,
   mockSecondaryTwoNoPicAndTwoSlice,
-  mockList2AndSixNoPicSlice
+  mockList2AndSixNoPicSlice,
+  mockLeadersSlice
 } from "@times-components/fixture-generator";
 import leadOneAndOneDataGenerator from "../fixtures/leadoneandone";
 import {
@@ -14,6 +15,7 @@ import {
   LeadOneFullWidthSlice,
   LeadTwoNoPicAndTwoSlice,
   ListTwoAndSixNoPicSlice,
+  LeadersSlice,
   SecondaryOneSlice,
   SecondaryFourSlice,
   SecondaryTwoAndTwoSlice,
@@ -170,6 +172,21 @@ export default () => {
           />
         );
 
+        expect(output).toMatchSnapshot();
+      }
+    },
+    {
+      name: "Leaders",
+      test: () => {
+        const leadersData = mockLeadersSlice();
+        const output = TestRenderer.create(
+          <LeadersSlice
+            leader1={leadersData.leader1}
+            leader2={leadersData.leader2}
+            leader3={leadersData.leader3}
+            onPress={() => {}}
+          />
+        );
         expect(output).toMatchSnapshot();
       }
     }