+++ packages/edition-slices/src/slices/leaders/masthead.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Image } from "react-native";
+import styles from "./styles";
+
+const MastHead = () => (
+  <Image
+    resizeMode="contain"
+    // eslint-disable-next-line global-require
+    source={require("../../../assets/leaders-masthead.png")}
+    style={styles.mastheadStyle}
+  />
+);
+
+export default MastHead;