+++ packages/edition-slices/src/tiles/tile-m/index.js
@@ -1,0 +1,33 @@
+import React from "react";
+import { View } from "react-native";
+import PropTypes from "prop-types";
+import { TileLink, TileSummary } from "../shared";
+import styles from "./styles";
+
+const TileM = ({ onPress, tile }) => {
+  const {
+    strapline,
+    article: { shortHeadline }
+  } = tile;
+  const tileWithoutLabelAndFlags = { article: { shortHeadline } };
+
+  return (
+    <TileLink onPress={onPress} tile={tileWithoutLabelAndFlags}>
+      <View style={styles.container}>
+        <TileSummary
+          headlineStyle={styles.headline}
+          strapline={strapline}
+          straplineStyle={styles.strapline}
+          tile={tileWithoutLabelAndFlags}
+        />
+      </View>
+    </TileLink>
+  );
+};
+
+TileM.propTypes = {
+  onPress: PropTypes.func.isRequired,
+  tile: PropTypes.shape({}).isRequired
+};
+
+export default TileM;