+++ packages/provider-queries/src/provider-queries.js
@@ -4,6 +4,7 @@ import * as authorArticlesNoImagesQuery from "./author-articles-no-images";
 import * as authorArticlesWithImagesQuery from "./author-articles-with-images";
 import authorQuery from "./author";
 import editionQuery from "./edition";
+import nativeEditionQuery from "./native-edition";
 import topicQuery from "./topic";
 import * as topicArticlesQuery from "./topic-articles";
 
@@ -17,6 +18,7 @@ export const authorArticlesWithImagesPTV =
   authorArticlesWithImagesQuery.propsToVariables;
 export const author = authorQuery;
 export const edition = editionQuery;
+export const nativeEdition = nativeEditionQuery;
 export const topic = topicQuery;
 export const topicArticles = topicArticlesQuery.default;
 export const topicArticlesPTV = topicArticlesQuery.propsToVariables;