+++ packages/slice-layout/__tests__/leaders.web.js
@@ -1,0 +1,5 @@
+import { mount } from "enzyme";
+import "./serializers.web";
+import shared from "./leaders.base";
+
+export default () => shared(mount);