+++ packages/slice-layout/src/templates/leaders/styles.js
@@ -1,0 +1,9 @@
+import { spacing } from "@times-components/styleguide";
+
+const styles = {
+  container: {
+    paddingBottom: spacing(1)
+  }
+};
+
+export default styles;