+++ ios-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.0...@times-components/ios-app@0.7.1) (2019-02-13)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 # [0.7.0](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.6.31...@times-components/ios-app@0.7.0) (2019-02-13)