+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.12.0...@times-components/article-main-comment@0.12.1) (2019-02-13)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 # [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@0.11.23...@times-components/article-main-comment@0.12.0) (2019-02-13)